requirements.txt
pandas
Pillow
streamlit
qrcode